Began porting to WebExtension API.
We'll need to add a popup because a background script alone can't write to the clipboard. The popup could do more useful things than just to list the tabs.

We could also try to do this via a content script, but they are blocked on certain domains, so we would have to default back to the popup in these domains anyhow.